fix for external monitor
changed "grep" --> "grep -m 1" in order to return only the first match. The modification prevent the script to fail when you have plugged in an external monitor. Otherwise, when you have an additional display connected, grep return two (or more) lines with the brightnesses of the two displays (external and the laptop's one).